Project-2/requirements.txt
torch
torchvision
torchsummary
opencv-python
timm
tqdm
numpy
matplotlib.pyplot
random